Refactor example.
